pfblockerng: enums, short-circuit, str-over-re in pfb_unbound.py (ADR-28 P8)
Inline pfb_cfg_adapters.py into pfb_unbound.py and delete the standalone file.
pfb_cfg_adapters.py was absent from the FreeBSD-ports pkg-plist, breaking every
.pkg build. pfb_unbound.py is the only consumer and the sole shipped file, so no
separate plist entry is needed or present.

Changes:
- Add `from enum import Enum` to pfb_unbound.py stdlib imports.
- Move IdnMode, PfbToggle, PfbLenient enums and the six pfb_cfg_*_read/write
  adapter helpers from pfb_cfg_adapters.py into pfb_unbound.py, placed after
  the IDN_MODE_OFF/ALL/CONFUSABLE constants. All docstrings preserved.
- Delete src/usr/local/pkg/pfblockerng/pfb_cfg_adapters.py (git rm).
- Repoint tests/test_adr28_cfg_adapters.py: import from pfb_unbound instead of
  pfb_cfg_adapters; drop now-redundant sys.path/Path hacks (conftest covers it).
  All 41 round-trip tests pass unchanged in intent and assertion.
- Update CLAUDE.md Python-adapter location line: pfb_cfg_adapters.py ->
  pfb_unbound.py.

Convention work (str-over-re, short-circuit, enum adoption at call sites):
- re->str swaps: zero. Every re use in pfb_unbound.py is a genuine pattern
  (compiled ReDoS detectors, user-rule patterns, multi-feature anchored classes).
  The is_idn_domain() hot path already uses str methods.
- Short-circuit reorders: zero. The idn_mode_decision() path already places the
  cheapest guard (is_idn_domain -- two str ops) first.
- Enum adoption at call sites: deferred. Threading IdnMode through the pfb dict
  / ini-read boundary is a non-trivial wiring change with regression risk in the
  ADR-08 Confusable/All arms (ADR-28 §2.4 boundary rule). Enums are defined and
  available; a future phase can adopt them at the read boundary safely.

Manifest/ini string contract intact. ADR-08 homoglyph analyzer untouched.
1691 tests pass; ruff/mypy clean; markdownlint 0 errors.
